Separate OpenAPI 3 oneOf validator
