+++ src/main/java/io/vertx/ext/json/schema/common/OneOfValidatorFactory.java
@@ -11,8 +11,6 @@
 package io.vertx.ext.json.schema.common;
 
 import io.vertx.core.Future;
-import io.vertx.core.json.JsonObject;
-import io.vertx.core.json.pointer.JsonPointer;
 import io.vertx.ext.json.schema.NoSyncValidationException;
 import io.vertx.ext.json.schema.ValidationException;
 
@@ -20,40 +18,20 @@ import java.util.Arrays;
 import java.util.stream.Collectors;
 
 import static io.vertx.ext.json.schema.ValidationException.createException;
-import java.util.List;
-import java.util.Map;
 
 public class OneOfValidatorFactory extends BaseCombinatorsValidatorFactory {
 
-    @Override
-    public Validator createValidator(JsonObject schema, JsonPointer scope, SchemaParserInternal parser, MutableStateValidator parent) {
-      OneOfValidator validator = (OneOfValidator) super.createValidator(schema, scope, parser, parent);
-      if (schema.containsKey("discriminator") && schema.getValue("discriminator") instanceof JsonObject) {
-        JsonObject discriminator = schema.getJsonObject("discriminator");
-        validator.setPropertyName(discriminator.getString("propertyName"));
-        if (discriminator.containsKey("mapping")) {
-          validator.setMapping(discriminator.getJsonObject("mapping").getMap().entrySet().stream()
-                    .collect(Collectors.toMap(entry -> entry.getKey(), entry -> entry.getValue().toString()))
-                );
-        }
-      }
-      return validator;
-    }
-
   @Override
-  BaseCombinatorsValidator instantiate(MutableStateValidator parent) {
+  protected BaseCombinatorsValidator instantiate(MutableStateValidator parent) {
     return new OneOfValidator(parent);
   }
 
   @Override
-  String getKeyword() {
+  protected String getKeyword() {
     return "oneOf";
   }
 
   class OneOfValidator extends BaseCombinatorsValidator {
-
-    private String propertyName;
-    private Map<String, String> mapping;
 
     public OneOfValidator(MutableStateValidator parent) {
       super(parent);
@@ -71,7 +49,7 @@ public class OneOfValidatorFactory extends BaseCombinatorsValidatorFactory {
     @Override
     public void validateSync(ValidatorContext context, Object in) throws ValidationException, NoSyncValidationException {
       this.checkSync();
-      long validCount = Arrays.stream(schemas).filter(s -> checkDiscriminator(s, in)).map(s -> isValidSync(s, context, in)).filter(b -> b.equals(true)).count();
+      long validCount = Arrays.stream(schemas).map(s -> isValidSync(s, context, in)).filter(b -> b.equals(true)).count();
       if (validCount > 1) throw createException("More than one schema valid", "oneOf", in);
       else if (validCount == 0) throw createException("No schema matches", "oneOf", in);
     }
@@ -79,34 +57,10 @@ public class OneOfValidatorFactory extends BaseCombinatorsValidatorFactory {
     @Override
     public Future<Void> validateAsync(ValidatorContext context, Object in) {
       if (isSync()) return validateSyncAsAsync(context, in);
-      List<Future<Void>> futures = Arrays.stream(schemas).filter(s -> checkDiscriminator(s, in)).map(s -> s.validateAsync(context, in)).collect(Collectors.toList());
-      if (futures.isEmpty()) {
-        return Future.failedFuture(createException("No schema matches", "oneOf", in));
-      } else {
-        return FutureUtils.oneOf(futures).recover(err -> Future.failedFuture(createException("No schema matches", "oneOf", in, err)));
-      }
-    }
-
-    private boolean checkDiscriminator(SchemaInternal schema, Object in) {
-      if (propertyName != null) {
-        if (in instanceof JsonObject) {
-          String discriminator = ((JsonObject) in).getString(propertyName);
-          if (mapping != null) discriminator = mapping.get(discriminator);
-
-          JsonObject jsonSchema = ((JsonObject) schema.getJson());
-          String schemaName = (jsonSchema.containsKey("name")) ? jsonSchema.getString("name") : jsonSchema.getString("$ref");
-          return schemaName!=null && discriminator != null && (discriminator.equals(schemaName) || schemaName.endsWith("/"+discriminator));
-        }
-      }
-      return true;
-    }
-
-    public void setPropertyName(String propertyName) {
-      this.propertyName = propertyName;
-    }
-
-    public void setMapping(Map<String, String> mapping) {
-      this.mapping = mapping;
+      return FutureUtils
+        .oneOf(Arrays.stream(schemas).map(s -> s.validateAsync(context, in)).collect(Collectors.toList()))
+        .recover(err -> Future.failedFuture(createException("No schema matches", "oneOf", in, err)));
     }
   }
+
 }
